Further refactoring:
* Use DependencyDeprecationDate object instead of passing env vars through the program
* Rename to BuildpackToml, and add methods to it
[#169324300]
